move test_get_blocks_from_canonical_chain_by_slot and test_get_blocks_from_fork_chain_by_root to test_utils
