+++ tests/libp2p/bcc/test_utils.py
@@ -1,6 +1,7 @@
 import asyncio
 from typing import NamedTuple
 
+from eth.exceptions import BlockNotFound
 from eth_keys import datatypes
 from libp2p.peer.id import ID
 import pytest
@@ -14,12 +15,15 @@ from trinity.protocol.bcc_libp2p.exceptions import (
 )
 from trinity.protocol.bcc_libp2p.messages import HelloRequest
 from trinity.protocol.bcc_libp2p.utils import (
+    get_blocks_from_canonical_chain_by_slot,
+    get_blocks_from_fork_chain_by_root,
     peer_id_from_pubkey,
     read_req,
     read_resp,
     write_req,
     write_resp,
 )
+from trinity.tools.bcc_factories import BeaconBlockFactory
 
 # Wrong type of `fork_version`, which should be `bytes4`.
 invalid_ssz_msg = HelloRequest(fork_version="1")
@@ -167,3 +171,74 @@ async def test_write_resp_failure():
     # Test: Raise `WriteMessageFailure` if `ssz.SerializationError` is thrown.
     with pytest.raises(WriteMessageFailure):
         await write_req(s, invalid_ssz_msg)
+
+
+@pytest.mark.parametrize(
+    "db_block_slots, slot_of_requested_blocks, expected_block_slots",
+    (
+        (range(5), [0, 2, 4], [0, 2, 4]),
+        ([1, 3, 5], [0, 2, 4], []),
+        ([2, 4], range(5), [2, 4]),
+    ),
+)
+@pytest.mark.asyncio
+async def test_get_blocks_from_canonical_chain_by_slot(
+    db_block_slots, slot_of_requested_blocks, expected_block_slots
+):
+    chain = BeaconBlockFactory.create_branch_by_slots(db_block_slots)
+    # Mock up block database
+    mock_slot_to_block_db = {block.slot: block for block in chain}
+
+    class Chain:
+        def get_canonical_block_by_slot(self, slot):
+            if slot in mock_slot_to_block_db:
+                return mock_slot_to_block_db[slot]
+            else:
+                raise BlockNotFound
+
+    result_blocks = get_blocks_from_canonical_chain_by_slot(
+        chain=Chain(), slot_of_requested_blocks=slot_of_requested_blocks
+    )
+
+    expected_blocks = [mock_slot_to_block_db[slot] for slot in expected_block_slots]
+    assert len(result_blocks) == len(expected_blocks)
+    assert set(result_blocks) == set(expected_blocks)
+
+
+@pytest.mark.parametrize(
+    "fork_chain_block_slots, slot_of_requested_blocks, expected_block_slots",
+    (
+        (range(10), [1, 4], [1, 4]),
+        ([0, 2, 3, 7, 8], list(range(1, 9, 2)), [3, 7]),
+        ([0, 2, 5], list(range(1, 6)), [2, 5]),
+        ([0, 4, 5], [2, 3], []),
+    ),
+)
+@pytest.mark.asyncio
+async def test_get_blocks_from_fork_chain_by_root(
+    fork_chain_block_slots, slot_of_requested_blocks, expected_block_slots
+):
+    fork_chain_blocks = BeaconBlockFactory.create_branch_by_slots(
+        fork_chain_block_slots
+    )
+    mock_root_to_block_db = {block.signing_root: block for block in fork_chain_blocks}
+
+    class Chain:
+        def get_block_by_root(self, root):
+            if root in mock_root_to_block_db:
+                return mock_root_to_block_db[root]
+            else:
+                raise BlockNotFound
+
+    requested_blocks = get_blocks_from_fork_chain_by_root(
+        chain=Chain(),
+        start_slot=slot_of_requested_blocks[0],
+        peer_head_block=fork_chain_blocks[-1],
+        slot_of_requested_blocks=slot_of_requested_blocks,
+    )
+
+    expected_blocks = [
+        block for block in fork_chain_blocks if block.slot in expected_block_slots
+    ]
+    assert len(requested_blocks) == len(expected_blocks)
+    assert set(requested_blocks) == set(expected_blocks)
